Use sparse matrices
for mass matrices and the matrix A in the field equations. In some
places we can use scipy.sparse.coo_arrays, but that doesn't work well
with numba, so in some places we just wrangle three numpy arrays for
lines, columns and values for each sparse matrix. That is a bit of a
pain, especially for the 9 mass matrices but works in the end.